fix some warnings from clippy
