+++ arrow-protocol/src/serde/ser.rs
@@ -22,6 +22,12 @@ impl Serializer {
     }
 }
 
+impl Default for Serializer {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 impl<'a> serde::Serializer for &'a mut Serializer {
     type Ok = ();
